requirements.txt
fastapi
google.generativeai
dotenv
pymongo
transformers
spacy
reportlab